Apply refactoring and improvements to Lua helpers

+++ src/Common/Filesystem.lua
@@ -17,7 +17,7 @@ function module.Exists(path)
 end
 
 function module.IsFolder(path)
-    if lfs.attributes(path, "mode") ~= "folder" then
+    if lfs.attributes(path, "mode") ~= "directory" then
         return false
     end
     return true
@@ -42,6 +42,8 @@ function module.GetCurrentFolder()
     return currentFolder
 end
 
+module.CurrentFolder = module.GetCurrentFolder()
+
 function module.CreateFolder(path)
     if module.Exists(path) then
         return
@@ -51,6 +53,30 @@ function module.CreateFolder(path)
     if result then
         console.PrintColor(path .. " created", console.Colors.Green)
     end
+end
+
+function module.IterateFolder(folder)
+    assert(folder and folder ~= "", "Parameter is missing or empty")
+    if string.sub(folder, -1) == "/" then
+        folder = string.sub(folder, 1, -2)
+    end
+
+    local function YieldFolder(folder)
+        for entry in lfs.dir(folder) do
+            if entry ~= "." and entry ~= ".." then
+                entry = folder .. "/" .. entry
+                local attribute = lfs.attributes(entry)
+                coroutine.yield(entry, attribute)
+                if attribute.mode == "directory" then
+                    YieldFolder(entry)
+                end
+            end
+        end
+    end
+
+    return coroutine.wrap(function()
+        YieldFolder(folder)
+    end)
 end
 
 function module.CopyFile(source, destination)
@@ -73,6 +99,10 @@ function module.CopyFile(source, destination)
 end
 
 function module.MoveFile(source, destination)
+    if not module.Exists(source) then
+        return
+    end
+
     local result, error = os.rename(source, destination)
     if not result then
         console.PrintColor("Error " .. error, console.Colors.Red)
@@ -80,10 +110,45 @@ function module.MoveFile(source, destination)
 end
 
 function module.DeleteFile(path)
+    if not module.IsFile(path) or not module.Exists(path) then
+        return
+    end
+
     local result, error = os.remove(path)
     if not result then
-        return console.PrintColor("Error " .. error, console.Colors.Red)
+        console.PrintColor("Error " .. error, console.Colors.Red)
     end
+end
+
+function module.DeleteFolder(path)
+    if not module.IsFolder(path) or not module.Exists(path) then
+        return
+    end
+
+    local result, error = lfs.rmdir(path)
+    if not result then
+        console.PrintColor("Error " .. error, console.Colors.Red)
+    end
+end
+
+function module.DeleteFolderRecursively(path)
+    if not module.IsFolder(path) or not module.Exists(path) then
+        return
+    end
+
+    for file, _ in module.IterateFolder(path) do
+        if module.IsFile(file) then
+            module.DeleteFile(file)
+        end
+    end
+
+    for folder, _ in module.IterateFolder(path) do
+        if module.IsFolder(folder) then
+            module.DeleteFolder(folder)
+        end
+    end
+
+    module.DeleteFolder(path)
 end
 
 module.TreeFolders = {
